# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"bearer":[]}]}},"/api/user/{id}":{"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditUserDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDto"}}}},"responses":{"201":{"description":""}}}},"/api/order":{"post":{"operationId":"OrderController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}}}},"get":{"operationId":"OrderController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}}},"security":[{"bearer":[]}]}},"/api/order/{uuid}":{"get":{"operationId":"OrderController_findOne","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}},"patch":{"operationId":"OrderController_update","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"delete":{"operationId":"OrderController_remove","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/api/product":{"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditProductDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"security":[{"bearer":[]}]},"get":{"operationId":"ProductController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}},"security":[{"bearer":[]}]}},"/api/product/filtered":{"get":{"operationId":"ProductController_foundWithFilter","parameters":[{"name":"syrupType","required":false,"in":"query","schema":{"enum":["AMBER","DARK","CLEAR"],"type":"string"}},{"name":"skip","required":false,"in":"query","schema":{"type":"string"}},{"name":"take","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/api/product/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditProductDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/api/product/{id}/refill":{"patch":{"operationId":"ProductController_refillStock","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"security":[{"bearer":[]}]}}},"info":{"title":"sugar-shark-api","description":"The sugar-shark-api API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"EditUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"}},"required":["email","password","firstName"]},"User":{"type":"object","properties":{}},"AuthUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"OrderLineDto":{"type":"object","properties":{"id":{"type":"number"},"productId":{"type":"number"},"productName":{"type":"string"},"productImage":{"type":"string"},"productPrice":{"type":"number"},"quantity":{"type":"number"},"syrupType":{"type":"string","enum":["AMBER","DARK","CLEAR"]}}},"OrderDto":{"type":"object","properties":{"uuid":{"type":"string"},"customerName":{"type":"string"},"customerEmail":{"type":"string"},"totalPrice":{"type":"number"},"orderStatus":{"enum":["creating","pending","created","shipped","delivered","cancelled"],"type":"string"},"orderLines":{"type":"array","items":{"$ref":"#/components/schemas/OrderLineDto"}}}},"Order":{"type":"object","properties":{}},"EditProductDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"image":{"type":"string"},"price":{"type":"number"},"stock":{"type":"number"},"maxStock":{"type":"number"},"syrupType":{"enum":["AMBER","DARK","CLEAR"],"type":"string"}}},"Product":{"type":"object","properties":{}}}}}